Deduplicate Dav1dLoopRestorationDSPContext
